% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimx.R
\name{fufn}
\alias{fufn}
\title{Test Function Data for use with Optimx}
\usage{
fufn(fnum)
}
\arguments{
\item{fnum}{Function number (1-35) to return data for.}
}
\description{
This function provides formatted data about each of the functions in
this package for ease of use with the \code{optimx} package.
}
\seealso{
\code{\link{fufnrun}} to run optimx using this data.
}
