% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimx.R
\name{fufnrun}
\alias{fufnrun}
\title{Run Optimx Using RFO Data}
\usage{
fufnrun(filename = "RFO.txt")
}
\arguments{
\item{filename}{path to an RFO file contain information on the functions and
methods from optimx to use.}
}
\description{
Use the information in an RFO file to run methods from optimx on the
functions in this package and report the results.
}
