% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_bard.R
\name{bard}
\alias{bard}
\title{Bard Function}
\usage{
bard()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 8 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 3}, \code{m = 15}.
  \item Minima: \code{f = 8.21487...e-3} and
  \code{f = 17.4286} at \code{(0.8406, -Inf, -Inf)}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Bard, Y. (1970).
Comparison of gradient methods for the solution of nonlinear parameter
estimation problems.
\emph{SIAM Journal on Numerical Analysis}, \emph{7}(1), 157-186.
\url{http://dx.doi.org/10.1137/0707011}
}
