% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_beale.R
\name{beale}
\alias{beale}
\title{Beale Function}
\usage{
beale()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 5 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 2}, \code{m = 3}.
  \item Minima: \code{f = 0} at \code{(3, 0.5) },
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Beale, E. M. L. (1958).
\emph{On an iterative method for finding a local minimum of a function of more
than one variable} (No. 25).
Statistical Techniques Research Group, Section of Mathematical Statistics,
Department of Mathematics, Princeton University.
}
