% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_box_3d.R
\name{box_3d}
\alias{box_3d}
\title{Box Three-Dimensional Function}
\usage{
box_3d(m = 20)
}
\arguments{
\item{m}{Number of summand functions in the objective function. Should be
equal to or greater than 3.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{he} If available, the hessian matrix (second derivatives)
  of the function w.r.t. the parameters at the given values.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 12 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{n = 3}, number of summand
  functions \code{m >= n}.
  \item Minima: \code{f = 0} at \code{(1, 10, 1), (10, 1, -1)} and
    where \code{x1 = x2} and \code{x3 = 0}.
}
}
\examples{
# Use 10 summand functions
fun <- box_3d(m = 10)
# Optimize using the standard starting point
x0 <- fun$x0
res_x0 <- stats::optim(par = x0, fn = fun$fn, gr = fun$gr, method =
"L-BFGS-B")
# Use your own starting point
res <- stats::optim(c(0.1, 0.2, 0.3), fun$fn, fun$gr, method = "L-BFGS-B")

# Use 20 summand functions
fun20 <- box_3d(m = 20)
res <- stats::optim(fun20$x0, fun20$fn, fun20$gr, method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{doi.org/10.1145/355934.355936}

Box, M. J. (1966).
A comparison of several current optimization methods, and the use of
transformations in constrained problems.
\emph{The Computer Journal}, \emph{9}(1), 67-77.
\doi{doi.org/10.1093/comjnl/9.1.67}
}
