% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/27_brown_al.R
\name{brown_al}
\alias{brown_al}
\title{Brown Almost-Linear Function}
\usage{
brown_al()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} Function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 27 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable, \code{m = n}.
  \item Minima: \code{f = 0} at \code{(a, a, a, ..., a ^ (1 - n))}, where
  \code{a} satisfies \code{n * a ^ n - (n + 1) * a ^ (n - 1) + 1 = 0};
  \code{f = 1} at \code{c(0, 0, ..., n + 1)}.
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
bal <- brown_al()
# 6 variable problem using the standard starting point
x0_6 <- bal$x0(6)
res_6 <- stats::optim(x0_6, bal$fn, bal$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(bal$x0(8), bal$fn, bal$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), bal$fn, bal$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Brown, K. M. (1969).
A quadratically convergent Newton-like method based upon Gaussian
elimination.
\emph{SIAM Journal on Numerical Analysis}, \emph{6}(4), 560-569.
\url{http://dx.doi.org/10.1137/0706051}
}
