% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_brown_bs.R
\name{brown_bs}
\alias{brown_bs}
\title{Brown Badly Scaled Function}
\usage{
brown_bs()
}
\value{
A list containing:
\itemize{
  \item \code{fn} The function.
  \item \code{gr} The gradient.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 4 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 2}, \code{m = 3}.
  \item Minima: \code{f = 0} at \code{(1e6, 2e-6) },
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}
}
