% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_brown_den.R
\name{brown_den}
\alias{brown_den}
\title{Brown and Dennis Function}
\usage{
brown_den(m = 20)
}
\arguments{
\item{m}{Number of terms in the objective function. Should be equal to or
greater than \code{n}.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 16 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 4}, \code{m >= n}.
  \item Minima: \code{f = 85822.2} if \code{m = 20}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Brown, K. M., & Dennis, J. E. (1971).
\emph{New computational algorithms for minimizing a sum of squares of
nonlinear functions} (Report No. 71-6).
New Haven, CT: Department of Computer Science, Yale University.
}
