% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_brown_den.R
\name{brown_den}
\alias{brown_den}
\title{Brown and Dennis Function}
\usage{
brown_den(m = 20)
}
\arguments{
\item{m}{Number of summand functions in the objective function. Should be
equal to or greater than 4.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 16 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{n = 4}, number of summand
  functions \code{m >= n}.
  \item Minima: \code{f = 85822.2} if \code{m = 20}.
}
}
\examples{
# Use 10 summand functions
fun <- brown_den(m = 10)
# Optimize using the standard starting point
x0 <- fun$x0
res_x0 <- stats::optim(par = x0, fn = fun$fn, gr = fun$gr, method =
"L-BFGS-B")
# Use your own starting point
res <- stats::optim(c(0.1, 0.2, 0.3, 0.4), fun$fn, fun$gr, method =
"L-BFGS-B")

# Use 20 summand functions
fun20 <- brown_den(m = 20)
res <- stats::optim(fun20$x0, fun20$fn, fun20$gr, method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Brown, K. M., & Dennis, J. E. (1971).
\emph{New computational algorithms for minimizing a sum of squares of
nonlinear functions} (Report No. 71-6).
New Haven, CT: Department of Computer Science, Yale University.
}
