% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/29_disc_ie.R
\name{disc_ie}
\alias{disc_ie}
\title{Discrete Integral Equation Function}
\usage{
disc_ie()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 29 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable, \code{m = n}.
  \item Minima: \code{f = 0} (at the same location as
  \code{\link{disc_bv}}).
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
d_ie <- disc_ie()
# 6 variable problem using the standard starting point
x0_6 <- d_ie$x0(6)
res_6 <- stats::optim(x0_6, d_ie$fn, d_ie$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(d_ie$x0(8), d_ie$fn, d_ie$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), d_ie$fn, d_ie$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

More', J. J., & Cosnard, M. Y. (1979).
Numerical solution of nonlinear equations.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{5}(1), 64-85.
\url{https://doi.org/10.1145/355815.355820}
}
