% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_ex_powell.R
\name{ex_powell}
\alias{ex_powell}
\title{Extended Powell Function}
\usage{
ex_powell()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 22 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable but a multiple of 4, \code{m = n}.
  \item Minima: \code{f = 0} at \code{rep(0, n)}
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
expow <- ex_powell()
# 12 variable problem using the standard starting point
x0_12 <- expow$x0(12)
res_12 <- stats::optim(x0_12, expow$fn, expow$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(expow$x0(8), expow$fn, expow$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), expow$fn, expow$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Spedicato, E. (1975).
\emph{Computational experience with quasi-Newton algorithms for minimization
problems of moderately large size} (Report CISE-N-175).
Segrate, Milano: Computing Center, CISE.
}
