% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_ex_rosen.R
\name{ex_rosen}
\alias{ex_rosen}
\title{Extended Rosenbrock Function}
\usage{
ex_rosen()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} Function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 21 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable but even, \code{m = n}.
  \item Minima: \code{f = 0} at \code{rep(1, n)}
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
exros <- ex_rosen()
# 6 variable problem using the standard starting point
x0_6 <- exros$x0(6)
res_6 <- stats::optim(x0_6, exros$fn, exros$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(exros$x0(8), exros$fn, exros$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), exros$fn, exros$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Spedicato, E. (1975).
\emph{Computational experience with quasi-Newton algorithms for minimization
problems of moderately large size} (Report CISE-N-175).
Segrate, Milano: Computing Center, CISE.
}
