% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_freudenstein_roth.R
\name{freud_roth}
\alias{freud_roth}
\title{Freudenstein and Roth Function}
\usage{
freud_roth()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 2 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 2}, \code{m = 2}.
  \item Minima: \code{f = 0} at \code{(5, 4)},
  \code{f = 48.9842...} at \code{(11.41..., -0.8968...)}
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Freudenstein, F., & Roth, B. (1963).
Numerical solution of systems of nonlinear equations.
\emph{Journal of the ACM (JACM)}, \emph{10}(4), 550-556.
\url{https://doi.org/10.1145/321186.321200}
}
