% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_helical_valley.R
\name{helical}
\alias{helical}
\title{Helical Valley Function}
\usage{
helical()
}
\value{
A list containing:
\itemize{
  \item \code{fn} The function.
  \item \code{gr} The gradient.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 7 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 3}, \code{m = 3}.
  \item Minima: \code{f = 0} at \code{(1, 0, 0)}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Fletcher, R., & Powell, M. J. (1963).
A rapidly convergent descent method for minimization.
\emph{The Computer Journal}, \emph{6}(2), 163-168.
}
