% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_jenn_samp.R
\name{jenn_samp}
\alias{jenn_samp}
\title{Jennrich and Sampson Function}
\usage{
jenn_samp(m = 10)
}
\arguments{
\item{m}{Number of summand functions in the objective function. Should be
equal to or greater than 2.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{he} If available, the hessian matrix (second derivatives)
  of the function w.r.t. the parameters at the given values.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 6 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{n = 2}, number of summand
  functions \code{m >= n}.
  \item Minima: \code{f = 124.362...} at \code{(x1 = x2 = 0.2578)} for
  \code{m = 10},
}
}
\note{
This test problem isn't really unconstrained. \code{x1} must take
a value between \code{(-1, 1)}. Included for the sake of completeness.
}
\examples{
# Use m = 10 summand functions
fun <- jenn_samp(m = 10)
# Optimize using the standard starting point
# Set 'lower' and 'upper' parameter to constrain par[1]. Only works with
# L-BFGS-B.
x0 <- fun$x0
res_x0 <- stats::optim(par = x0, fn = fun$fn, gr = fun$gr, method =
"L-BFGS-B", lower = -1, upper = 1)
# Use your own starting point
res <- stats::optim(c(0.1, 0.2), fun$fn, fun$gr, method = "L-BFGS-B",
lower = -1, upper = 1)

# Use 20 summand functions
fun20 <- jenn_samp(m = 20)
res <- stats::optim(fun20$x0, fun20$fn, fun20$gr, method = "L-BFGS-B",
lower = -1, upper = 1)
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{doi.org/10.1145/355934.355936}

Jennrich, R. I., & Sampson, P. F. (1968).
Application of stepwise regression to non-linear estimation.
\emph{Technometrics}, \emph{10}(1), 63-72.
}
