% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_jenn_samp.R
\name{jenn_samp}
\alias{jenn_samp}
\title{Jennrich and Sampson Function}
\usage{
jenn_samp(m = 10)
}
\arguments{
\item{m}{Number of terms in the objective function.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 6 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 2}, \code{m >= n}.
  \item Minima: \code{f = 124.362...} at \code{(x1 = x2 = 0.2578)} for
  \code{m = 10},
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Jennrich, R. I., & Sampson, P. F. (1968).
Application of stepwise regression to non-linear estimation.
\emph{Technometrics}, \emph{10}(1), 63-72.
}
