% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_kow_osb.R
\name{kow_osb}
\alias{kow_osb}
\title{Kowalik and Osborne Function}
\usage{
kow_osb()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 15 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 4}, \code{m = 11}.
  \item Minima: \code{f = 3.07505...e-4};
  and \code{f = 1.02734...e-3} at \code{(Inf, -14.07..., -Inf, -Inf)}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Kowalik, J. S., & Osborne, M. R. (1968).
\emph{Methods for unconstrained optimization problems.}
New York, NY: Elsevier North-Holland.
}
