% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_linfun_fr.R
\name{linfun_fr}
\alias{linfun_fr}
\title{Linear Function - Full Rank}
\usage{
linfun_fr(m)
}
\arguments{
\item{m}{Number of terms in the objective function. Should be equal to or
greater than \code{n}.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 32 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable, \code{m >= n}.
  \item Minima: \code{f = m - n} at \code{rep(-1, n)}.
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
linfr <- linfun_fr(m = 10)
# 6 variable problem using the standard starting point
x0_6 <- linfr$x0(n = 6)
res_6 <- stats::optim(x0_6, linfr$fn, linfr$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(linfr$x0(8), linfr$fn, linfr$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), linfr$fn, linfr$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}
}
