% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32_linfun_fr.R
\name{linfun_fr}
\alias{linfun_fr}
\title{Linear Function - Full Rank}
\usage{
linfun_fr(m = 100)
}
\arguments{
\item{m}{Number of summand functions in the objective function. Should be
equal to or greater than \code{n}.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 32 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{n} variable, \code{m >= n}.
  \item Minima: \code{f = m - n} at \code{rep(-1, n)}.
}

The number of parameters, \code{n}, in the objective function is not
specified when invoking this function. It is implicitly set by the length of
the parameter vector passed to the objective and gradient functions that this
function creates. See the 'Examples' section.
}
\examples{
linfr <- linfun_fr(m = 10)
# 6 variable problem using the standard starting point
x0_6 <- linfr$x0(n = 6)
res_6 <- stats::optim(x0_6, linfr$fn, linfr$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(linfr$x0(8), linfr$fn, linfr$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), linfr$fn, linfr$gr,
                      method = "L-BFGS-B")

# Use 20 summand functions
linfr_m20 <- linfun_fr(m = 20)
# Repeat 4 parameter optimization with new test function
res_n4_m20 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), linfr_m20$fn, linfr_m20$gr,
method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}
}
