% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_linfun_r1.R
\name{linfun_r1}
\alias{linfun_r1}
\title{Linear Function - Rank 1}
\usage{
linfun_r1(m)
}
\arguments{
\item{m}{Number of terms in the objective function. Should be equal to or
greater than \code{n}.}
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 33 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable, \code{m >= n}.
  \item Minima: \code{f = (m * (m - 1)) / (2 * (2 * m + 1))} at any set of
  points \code{xj} with \code{j = 1, 2, ..., n} where the sum of
  \code{j * xj = 3 / (2 * m + 1)}.
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
linr1 <- linfun_r1(m = 10)
# 6 variable problem using the standard starting point
x0_6 <- linr1$x0(n = 6)
res_6 <- stats::optim(x0_6, linr1$fn, linr1$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(linr1$x0(8), linr1$fn, linr1$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), linr1$fn, linr1$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}
}
