% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_meyer.R
\name{meyer}
\alias{meyer}
\title{Meyer Function}
\usage{
meyer()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 10 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 3}, \code{m = 16}.
  \item Minima: \code{f = 87.9458...}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Meyer, R. R. (1970).
Theoretical and computational aspects of nonlinear regression.
In J. B. Rosen, O. L. Mangasarian, and K. Ritter (Eds.)
\emph{Nonlinear programming} (pp465-496).
New York: Academic Press.
}
