% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_osborne_1.R
\name{osborne_1}
\alias{osborne_1}
\title{Osborne 1 Function}
\usage{
osborne_1()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 17 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 5}, \code{m = 33}.
  \item Minima: \code{f = 5.46489...e-5}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Osborne, M. R. (1972).
Some aspects of nonlinear least squares calculations.
In F. A. Lootsma (Ed.),
\emph{Numerical methods for nonlinear optimization} (pp. 171-189).
New York, NY: Academic Press.
}
