% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_osborne_2.R
\name{osborne_2}
\alias{osborne_2}
\title{Osborne 2 Function}
\usage{
osborne_2()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 19 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 11}, \code{m = 65}.
  \item Minima: \code{f = 4.01377...e-2}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Osborne, M. R. (1972).
Some aspects of nonlinear least squares calculations.
In F. A. Lootsma (Ed.),
\emph{Numerical methods for nonlinear optimization} (pp. 171-189).
New York, NY: Academic Press.
}
