% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_penalty_1.R
\name{penalty_1}
\alias{penalty_1}
\title{Penalty Function I}
\usage{
penalty_1()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 23 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable, \code{m = n + 1}.
  \item Minima: \code{f = 2.24997...e-5} if \code{n = 4};
  \code{f = 7.08765...e-5} if \code{n = 10}.
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
pen1 <- penalty_1()
# 6 variable problem using the standard starting point
x0_6 <- pen1$x0(6)
res_6 <- stats::optim(x0_6, pen1$fn, pen1$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(pen1$x0(8), pen1$fn, pen1$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), pen1$fn, pen1$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Gill, P. E., Murray, W., & Pitfield, R. A. (1972).
\emph{The implementation of two revised quasi-Newton algorithms for
unconstrained optimization} (Report NAC-11).
Teddington, London: National Physical Laboratory, Division of Numerical
Analysis and Computing.
}
