% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_powell_bs.R
\name{powell_bs}
\alias{powell_bs}
\title{Powell Badly Scaled Function}
\usage{
powell_bs()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 3 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 2}, \code{m = 2}.
  \item Minima: \code{f = 0} at \code{(1.098... e-5, 9.106...) },
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Powell, M. J. D. (1970).
A hybrid method for nonlinear equations.
In P. Rabinowitz (Ed.),
\emph{Numerical Methods for Nonlinear Algebraic Equations} (pp87-114).
New York: Gordon & Breach.
}
