% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_powell_bs.R
\name{powell_bs}
\alias{powell_bs}
\title{Powell Badly Scaled Function}
\usage{
powell_bs()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{he} If available, the hessian matrix (second derivatives)
  of the function w.r.t. the parameters at the given values.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
  \item \code{fmin} reported minimum
  \item \code{xmin} parameters at reported minimum
}
}
\description{
Test function 3 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{n = 2}, number of summand
  functions \code{m = 2}.
  \item Minima: \code{f = 0} at \code{(1.098... e-5, 9.106...) },
}
}
\examples{
fun <- powell_bs()
# Optimize using the standard starting point
x0 <- fun$x0
res_x0 <- stats::optim(par = x0, fn = fun$fn, gr = fun$gr, method =
"L-BFGS-B")
# Use your own starting point
res <- stats::optim(c(0.1, 0.2), fun$fn, fun$gr, method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{doi.org/10.1145/355934.355936}

Powell, M. J. D. (1970).
A hybrid method for nonlinear equations.
In P. Rabinowitz (Ed.),
\emph{Numerical Methods for Nonlinear Algebraic Equations} (pp87-114).
New York: Gordon & Breach.
}
