% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_powell_s.R
\name{powell_s}
\alias{powell_s}
\title{Powell Singular Function}
\usage{
powell_s()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 13 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 4}, \code{m = 4}.
  \item Minima: \code{f = 0} at \code{rep(0, 4)}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Powell, M. J. D. (1962).
An iterative method for finding stationary values of a function of several
variables.
\emph{The Computer Journal}, \emph{5}(2), 147-151.
\url{https://doi.org/10.1093/comjnl/5.2.147}
}
