% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rosen.R
\name{rosen}
\alias{rosen}
\title{Rosenbrock Function}
\usage{
rosen()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 1 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{n = 2}, number of summand
  functions \code{m = 2}.
  \item Minima: \code{f = 0} at \code{(1, 1)}
}
}
\examples{
fun <- rosen()
# Optimize using the standard starting point
x0 <- fun$x0
res_x0 <- stats::optim(par = x0, fn = fun$fn, gr = fun$gr, method =
"L-BFGS-B")
# Use your own starting point
res <- stats::optim(c(0.1, 0.2), fun$fn, fun$gr, method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Rosenbrock, H. (1960).
An automatic method for finding the greatest or least value of a function.
\emph{The Computer Journal}, \emph{3}(3), 175-184.
\url{https://doi.org/10.1093/comjnl/3.3.175}
}
