% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_var_dim.R
\name{var_dim}
\alias{var_dim}
\title{Variably Dimensioned Function}
\usage{
var_dim()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} Function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 25 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n} variable, \code{m = n + 2}.
  \item Minima: \code{f = 0} at \code{rep(1, n)}.
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
vdim <- var_dim()
# 6 variable problem using the standard starting point
x0_6 <- vdim$x0(6)
res_6 <- stats::optim(x0_6, vdim$fn, vdim$gr, method = "L-BFGS-B")
# Standing starting point with 8 variables
res_8 <- stats::optim(vdim$x0(8), vdim$fn, vdim$gr, method = "L-BFGS-B")
# Create your own 4 variable starting point
res_4 <- stats::optim(c(0.1, 0.2, 0.3, 0.4), vdim$fn, vdim$gr,
                      method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}
}
