% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_watson.R
\name{watson}
\alias{watson}
\title{Watson Function}
\usage{
watson()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} Function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
}
}
\description{
Test function 20 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{2 <= n <= 31}, \code{m = 31}.
  \item Minima: \code{f = 2.28767...e-3} if \code{n = 6};
  \code{f = 1.39976...e-6} if \code{n = 9};
  \code{f = 4.72238...e-10} if \code{n = 12}
}

The number of variables in the function, \code{n}, is determined by the
length of the vector passed to the function and gradient routines. See
the 'Examples' section.
}
\examples{
wat <- watson()
# 6 variable problem using the standard starting point
x0_6 <- wat$x0(6)
res_6 <- stats::optim(x0_6, wat$fn, wat$gr, method = "L-BFGS-B")
# Standing starting point with 9 variables
x0_9 <- wat$x0(9)
res_9 <- stats::optim(x0_9, wat$fn, wat$gr, method = "L-BFGS-B")
# Create your own 3 variable starting point
res_3 <- stats::optim(c(0.1, 0.2, 0.3), wat$fn, wat$gr, method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Kowalik, J. S., & Osborne, M. R. (1968).
\emph{Methods for unconstrained optimization problems.}
New York, NY: Elsevier North-Holland.
}
