% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_watson.R
\name{watson}
\alias{watson}
\title{Watson Function}
\usage{
watson()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Objective function which calculates the value given input
  parameter vector.
  \item \code{gr} Gradient function which calculates the gradient vector
  given input parameter vector.
  \item \code{he} If available, the hessian matrix (second derivatives)
  of the function w.r.t. the parameters at the given values.
  \item \code{fg} A function which, given the parameter vector, calculates
  both the objective value and gradient, returning a list with members
  \code{fn} and \code{gr}, respectively.
  \item \code{x0} Function returning the standard starting point, given
  \code{n}, the number of variables desired.
  \item \code{fmin} reported minimum
  \item \code{xmin} parameters at reported minimum
}
}
\description{
Test function 20 from the More', Garbow and Hillstrom paper.
}
\details{
The objective function is the sum of \code{m} functions, each of \code{n}
parameters.

\itemize{
  \item Dimensions: Number of parameters \code{2 <= n <= 31}, number of
  summand functions \code{m = 31}.
  \item Minima: \code{f = 2.28767...e-3} if \code{n = 6};
  \code{f = 1.39976...e-6} if \code{n = 9};
  \code{f = 4.72238...e-10} if \code{n = 12}
}

The number of parameters, \code{n}, in the objective function is not
specified when invoking this function. It is implicitly set by the length of
the parameter vector passed to the objective and gradient functions that this
function creates. See the 'Examples' section.
}
\examples{
wat <- watson()
# 6 variable problem using the standard starting point
x0_6 <- wat$x0(6)
res_6 <- stats::optim(x0_6, wat$fn, wat$gr, method = "L-BFGS-B")
# Standing starting point with 9 variables
x0_9 <- wat$x0(9)
res_9 <- stats::optim(x0_9, wat$fn, wat$gr, method = "L-BFGS-B")
# Create your own 3 variable starting point
res_3 <- stats::optim(c(0.1, 0.2, 0.3), wat$fn, wat$gr, method = "L-BFGS-B")
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{doi.org/10.1145/355934.355936}

Kowalik, J. S., & Osborne, M. R. (1968).
\emph{Methods for unconstrained optimization problems.}
New York, NY: Elsevier North-Holland.
}
