% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_wood.R
\name{wood}
\alias{wood}
\title{Wood Function}
\usage{
wood()
}
\value{
A list containing:
\itemize{
  \item \code{fn} Function which calculates the value given input
  parameter vector.
  \item \code{gr} Function which calculates the gradient vector given input
  parameter vector.
  \item \code{fg} A function which calculates both the value and gradient,
  given the parameter vector, returning a list.
  \item \code{x0} Standard starting point.
}
}
\description{
Test function 14 from the More', Garbow and Hillstrom paper.
}
\details{
\itemize{
  \item Dimensions: \code{n = 4}, \code{m = 6}.
  \item Minima: \code{f = 0} at \code{rep(1, 4)}.
}
}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\url{https://doi.org/10.1145/355934.355936}

Colville, A. R. (1968)
A comparative study on nonlinear programming codes (Report No. 320-2949).
New York, NY: IBM.
}
